refactor: shared key storage code
