+++ core/service/src/vault/storage/crypto_material/mod.rs
@@ -46,8 +46,8 @@ mod tests;
 
 // Re-export the public API for external consumption
 pub use base::CryptoMaterialStorage;
+pub use base::PublicKeySet;
 pub use centralized::CentralizedCryptoMaterialStorage;
-pub use threshold::PublicKeySet;
 pub use threshold::ThresholdCryptoMaterialStorage;
 pub use traits::CryptoMaterialReader;
 pub use traits::PrivateCryptoMaterialReader;
